requirements.txt
fastapi==0.129.1
uvicorn==0.41.0
python-multipart==0.0.22
openai==2.21.0
python-dotenv==1.2.1
dspy==3.1.3
pypdf==6.7.2
pydantic==2.12.5
pytest==9.0.2
httpx==0.28.1